+++ src/main/java/com/jandi/plan_backend/commu/comment/repository/CommentRepository.java
@@ -5,6 +5,9 @@ import com.jandi.plan_backend.commu.community.entity.Community;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 import java.util.Optional;
@@ -37,4 +40,25 @@ public interface CommentRepository extends JpaRepository<Comment, Integer> {
     List<Comment> findByUserIdAndParentCommentIsNull(Integer userId);
 
     List<Comment> findByUserIdAndParentCommentIsNotNull(Integer userId);
+
+    /** 증감 쿼리 */
+    // 게시글 댓글 수 증가
+    @Modifying
+    @Query("update Community c set c.commentCount = c.commentCount + :count where c.postId = :postId")
+    void increaseCommentCount(@Param("postId") Integer postId, @Param("count") int count);
+
+    // 게시글 댓글 수 감소
+    @Modifying
+    @Query("update Community c set c.commentCount = c.commentCount - :count where c.postId = :postId")
+    void decreaseCommentCount(@Param("postId") Integer postId, @Param("count") int count);
+
+    // 부모 댓글의 답글 수 증가
+    @Modifying
+    @Query("update Comment c set c.repliesCount = c.repliesCount + 1 where c.commentId = :commentId")
+    void increaseRepliesCount(@Param("commentId") Integer commentId);
+
+    // 부모 댓글의 답글 수 감소
+    @Modifying
+    @Query("update Comment c set c.repliesCount = c.repliesCount - 1 where c.commentId = :commentId")
+    void decreaseRepliesCount(@Param("commentId") Integer commentId);
 }
